+++ tests/conftest.py
@@ -1,4 +1,4 @@
-""" Test suite shared objects and setup """
+"""Test suite shared objects and setup"""
 
 import os
 from glob import glob